refactor: split CLI out-format implementation
