+++ crates/cli/src/out_format/mod.rs
@@ -1,0 +1,9 @@
+#![deny(unsafe_code)]
+
+mod parser;
+mod render;
+mod tokens;
+
+pub(crate) use parser::parse_out_format;
+pub(crate) use render::emit_out_format;
+pub(crate) use tokens::{OutFormat, OutFormatContext};
